Add files via upload
Bitcoin was introduced in 2008, in the backdrop of a financial crisis that put millions of lives in unstable situations. Many blamed the central bank for its inability to forecast the recession and lost faith in a centralization system. This created a void of trust that Bitcoin was able to fill at the time. With its decentralization system, Bitcoin empowered individuals who join the network to oversee the system themselves, instead of relying on a specific group or member to ensure their well-beings within the ecosystem. Over the decade, Bitcoin has become a potential disruptive cryptocurrency within the economy that increasingly drew attention. While this prompted government to intervene and produce regulation, it also inspired many people to start their own ventures that contribute to a movement of decentralization. In this research project, we aim to describe and visualize the story behind the rise of Bitcoin and its Network, and also examine its current state.